requirements.txt
django==3.2
python-dotenv
PyMySQL>=1.0.0
whitenoise==5.3.0
requests
djangorestframework
django-cors-headers